app/requirements.txt
flasgger
Flask
flask_cors
flask_sqlalchemy
passlib
PyJWT
SQLAlchemy
python-dotenv
